Create README.md
We have the data for the 100 top-rated movies from the past decade along with various pieces of information about the movie, its actors, and the voters who have rated these movies online. In this, we will try to find some interesting insights into these movies and their voters, using Python.
In this project we have written Python code to explore the data, gain insights into the movies, actors, votes, ratings and collections.